requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
gradio
requests
beautifulsoup4
firecrawl